+++ Octokit/Clients/ICommitCommentReactionsClient.cs
@@ -20,7 +20,7 @@ namespace Octokit
         /// <param name="number">The comment id</param>
         /// <param name="reaction">The reaction to create</param>
         /// <returns></returns>
-        Task<Reaction> Create(string owner, string name, int number, NewReaction reaction);
+        Task<Reaction> Create(string owner, string name, long number, NewReaction reaction);
 
         /// <summary>
         /// Creates a reaction for a specified Commit Comment
@@ -30,7 +30,7 @@ namespace Octokit
         /// <param name="number">The comment id</param>
         /// <param name="reaction">The reaction to create</param>
         /// <returns></returns>
-        Task<Reaction> Create(long repositoryId, int number, NewReaction reaction);
+        Task<Reaction> Create(long repositoryId, long number, NewReaction reaction);
 
         /// <summary>
         /// Get all reactions for a specified Commit Comment
@@ -40,7 +40,7 @@ namespace Octokit
         /// <param name="name">The name of the repository</param>
         /// <param name="number">The comment id</param>        
         /// <returns></returns>
-        Task<IReadOnlyList<Reaction>> GetAll(string owner, string name, int number);
+        Task<IReadOnlyList<Reaction>> GetAll(string owner, string name, long number);
 
         /// <summary>
         /// Get all reactions for a specified Commit Comment
@@ -51,7 +51,7 @@ namespace Octokit
         /// <param name="number">The comment id</param>
         /// <param name="options">Options for changing the API response</param>
         /// <returns></returns>
-        Task<IReadOnlyList<Reaction>> GetAll(string owner, string name, int number, ApiOptions options);
+        Task<IReadOnlyList<Reaction>> GetAll(string owner, string name, long number, ApiOptions options);
 
         /// <summary>
         /// Get all reactions for a specified Commit Comment
@@ -60,7 +60,7 @@ namespace Octokit
         /// <param name="repositoryId">The owner of the repository</param>
         /// <param name="number">The comment id</param>        
         /// <returns></returns>
-        Task<IReadOnlyList<Reaction>> GetAll(long repositoryId, int number);
+        Task<IReadOnlyList<Reaction>> GetAll(long repositoryId, long number);
 
         /// <summary>
         /// Get all reactions for a specified Commit Comment
@@ -70,6 +70,6 @@ namespace Octokit
         /// <param name="number">The comment id</param>
         /// <param name="options">Options for changing the API response</param> 
         /// <returns></returns>
-        Task<IReadOnlyList<Reaction>> GetAll(long repositoryId, int number, ApiOptions options);
+        Task<IReadOnlyList<Reaction>> GetAll(long repositoryId, long number, ApiOptions options);
     }
 }
